requirements.txt
tqdm==4.66.1
pandas==2.1.1
requests==2.31.0
streamlit==1.27.2
python-dotenv==1.0.0
elasticsearch==7.15.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2